Extract GpxExport.ExportTask
