+++ main/src/main/java/cgeo/geocaching/export/GpxExport.java
@@ -1,22 +1,14 @@
 package cgeo.geocaching.export;
 
-import cgeo.geocaching.CgeoApplication;
 import cgeo.geocaching.R;
-import cgeo.geocaching.activity.ActivityMixin;
 import cgeo.geocaching.models.Geocache;
 import cgeo.geocaching.settings.Settings;
-import cgeo.geocaching.storage.ContentStorage;
 import cgeo.geocaching.storage.PersistableFolder;
 import cgeo.geocaching.ui.dialog.Dialogs;
-import cgeo.geocaching.utils.AsyncTaskWithProgress;
 import cgeo.geocaching.utils.FileNameCreator;
-import cgeo.geocaching.utils.Log;
-import cgeo.geocaching.utils.ShareUtils;
-import cgeo.geocaching.utils.UriUtils;
 
 import android.app.Activity;
 import android.app.Dialog;
-import android.net.Uri;
 import android.view.View;
 import android.widget.CheckBox;
 import android.widget.TextView;
@@ -25,16 +17,8 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
 
-import java.io.BufferedWriter;
-import java.io.IOException;
-import java.io.OutputStream;
-import java.io.OutputStreamWriter;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
-import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 
 public class GpxExport extends AbstractExport {
@@ -60,7 +44,7 @@ public class GpxExport extends AbstractExport {
         if (activity == null) {
             // No activity given, so no user interaction possible.
             // Start export with default parameters.
-            new ExportTask(null).execute(geocodes);
+            new GpxExportExportTask(null, getProgressTitle(), fileName, getName()).execute(geocodes);
 
         } else {
             // Show configuration dialog
@@ -103,7 +87,7 @@ public class GpxExport extends AbstractExport {
         builder.setPositiveButton(R.string.export, (dialog, which) -> {
             Settings.setIncludeFoundStatus(includeFoundStatus.isChecked());
             dialog.dismiss();
-            new ExportTask(activity).execute(geocodes);
+            new GpxExportExportTask(activity, getProgressTitle(), fileName, getName()).execute(geocodes);
         });
 
         return builder.create();
@@ -113,60 +97,4 @@ public class GpxExport extends AbstractExport {
         return Geocache.getGeocodes(caches).toArray(new String[caches.size()]);
     }
 
-    protected class ExportTask extends AsyncTaskWithProgress<String, Uri> {
-
-        /**
-         * Instantiates and configures the task for exporting field notes.
-         *
-         * @param activity optional: Show a progress bar and toasts
-         */
-        public ExportTask(final Activity activity) {
-            super(activity, getProgressTitle());
-        }
-
-        @Override
-        protected Uri doInBackgroundInternal(final String[] geocodes) {
-            final List<String> allGeocodes = new ArrayList<>(Arrays.asList(geocodes));
-
-            setMessage(CgeoApplication.getInstance().getResources().getQuantityString(R.plurals.cache_counts, allGeocodes.size(), allGeocodes.size()));
-
-            final Uri uri = ContentStorage.get().create(PersistableFolder.GPX, fileName);
-            if (uri == null) {
-                return null;
-            }
-
-            BufferedWriter writer = null;
-            try (OutputStream os = ContentStorage.get().openForWrite(uri)) {
-                if (os == null) {
-                    return null;
-                }
-
-                writer = new BufferedWriter(new OutputStreamWriter(os, StandardCharsets.UTF_8));
-                new GpxSerializer().writeGPX(allGeocodes, writer, ExportTask.this::publishProgress);
-            } catch (final IOException e) {
-                Log.e("GpxExport.ExportTask export", e);
-                // delete partial GPX file on error
-                ContentStorage.get().delete(uri);
-
-                return null;
-            } finally {
-                IOUtils.closeQuietly(writer);
-            }
-
-            return uri;
-        }
-
-        @Override
-        protected void onPostExecuteInternal(final Uri uri) {
-            final Activity activityLocal = activity;
-            if (activityLocal != null) {
-                if (uri != null) {
-                    ShareUtils.shareOrDismissDialog(activityLocal, uri, "application/xml", R.string.export, getName() + ' ' + activityLocal.getString(R.string.export_exportedto) + ": " + UriUtils.toUserDisplayableString(uri));
-                } else {
-                    ActivityMixin.showToast(activityLocal, activityLocal.getString(R.string.export_failed));
-                }
-            }
-        }
-
-    }
 }
